+++ src/utm.rs
@@ -7,16 +7,15 @@ use coord::Coord;
 ///
 /// # Examples
 /// ```
-/// use std::error::Error;
-///
 /// extern crate geomorph;
+/// use geomorph::*;
 ///
-/// fn try_main() -> Result<geomorph::utm::Utm, geomorph::ParseError> {
+/// fn try_main() -> Result<utm::Utm, ParseError> {
 ///     let lat: f64 = -23.0095839;
 ///     let lon: f64 = -43.4361816;
 ///
-///     let coord = geomorph::coord::Coord::new(lat, lon)?;
-///     geomorph::utm::Utm::new(coord)
+///     let coord = coord::Coord::new(&lat, &lon)?;
+///     utm::Utm::new(&coord)
 /// }
 ///
 /// fn main() {
@@ -38,7 +37,7 @@ impl Utm {
     ///
     /// Inspired on the work of Rafael Palacios.
     ///
-    pub fn new(coord: Coord) -> Result<Utm, ParseError> {
+    pub fn new(coord: &Coord) -> Result<Utm, ParseError> {
         let lat = coord.lat;
         let lon = coord.lon;
 
@@ -149,7 +148,7 @@ mod tests {
     #[test]
     fn utm_zone_south() {
         let coord = Coord {lat: -23.0095839, lon: -43.4361816};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 660265.0944068021).abs() < 1.0);
         assert!((utm.northing - 7454564.243324452).abs() < 1.0);
         assert_eq!(utm.north, false);
@@ -160,7 +159,7 @@ mod tests {
     #[test]
     fn utm_zone_north() {
         let coord = Coord {lat: 52.517153, lon: 13.412389};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 392273.6051633584).abs() < 1.0);
         assert!((utm.northing - 5819744.4599129185).abs() < 1.0);
         assert_eq!(utm.north, true);
@@ -171,7 +170,7 @@ mod tests {
     #[test]
     fn utm_norway_zone() {
         let coord = Coord {lat: 61.076521, lon: 4.680180};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 267038.76).abs() < 1.0);
         assert!((utm.northing - 6779002.66).abs() < 1.0);
         assert_eq!(utm.north, true);
@@ -182,7 +181,7 @@ mod tests {
     #[test]
     fn utm_svalbard_zone_1() {
         let coord = Coord {lat: 78.891608, lon: 10.457194};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 402386.73).abs() < 1.0);
         assert!((utm.northing - 8761675.98).abs() < 1.0);
         assert_eq!(utm.north, true);
@@ -193,7 +192,7 @@ mod tests {
     #[test]
     fn utm_svalbard_zone_2() {
         let coord = Coord {lat: 78.122200, lon: 20.349504};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 622751.81).abs() < 1.0);
         assert!((utm.northing - 8677619.41).abs() < 1.0);
         assert_eq!(utm.north, true);
@@ -204,7 +203,7 @@ mod tests {
     #[test]
     fn utm_svalbard_zone_3() {
         let coord = Coord {lat: 78.102575, lon: 21.013745};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 362459.56).abs() < 1.0);
         assert!((utm.northing - 8676854.75).abs() < 1.0);
         assert_eq!(utm.north, true);
@@ -215,11 +214,26 @@ mod tests {
     #[test]
     fn utm_svalbard_zone_4() {
         let coord = Coord {lat: 78.138264, lon: 30.194746};
-        let utm = Utm::new(coord).unwrap();
+        let utm = Utm::new(&coord).unwrap();
         assert!((utm.easting - 573272.89).abs() < 1.0);
         assert!((utm.northing - 8675799.74).abs() < 1.0);
         assert_eq!(utm.north, true);
         assert_eq!(utm.zone, 35);
         assert_eq!(utm.band, 'X');
+    }
+
+    #[test]
+    fn coords_borrow() {
+        let lat: f64 = -34.073088;
+        let lon: f64 = 18.549757;
+        let coord = Coord::new(&lat, &lon).unwrap();
+        let utm = Utm::new(&coord).unwrap();
+        assert!((utm.easting - 273893.5).abs() < 1.0);
+        assert!((utm.northing - 6227030.5).abs() < 1.0);
+        assert_eq!(utm.north, false);
+        assert_eq!(utm.zone, 34);
+        assert_eq!(utm.band, 'H');
+        assert_eq!(coord.lat, lat);
+        assert_eq!(coord.lon, lon);
     }
 }