+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
+## [Todo]
+
+* Implement MGRS
+* Replace ParseError for standard Error
+* Implement `Utm.to_coords()`
+* Improve readability of `Utm::new(coord: Coord)`
+* Improve precision of `Utm::new(coord: Coord)`
+
 ## [Unreleased]
+
+* Implement `Coord.to_utm()`
 
 ## [0.2.0] - 2018-06-06
 
@@ -9,7 +19,5 @@
 * Some doc links were fixed.
 
 ## [0.1.0] - 2018-06-06
-
-### Fixed
 
 * Initial version.